# assets/openapi.yaml
openapi: 3.1.0
info:
    title: Phishing API
    summary: Una base de datos con miles enlaces maliciosos.
    description: Una base de datos colaborativa con enlaces maliciosos centrada pero no dedicada en la moderación de Discord.
    termsOfService: 'https://discusss.gitbook.io/phishing/informacion'
    contact:
      name: 'Discusss'
      url: 'https://discord.gg/SJqrGP7'
      email: 'admin@lacabra.app'
    license:
      name: GPL-3.0
      url: 'https://www.gnu.org/licenses/gpl-3.0.html'
    version: 0.0.1-dev
...
servers:
  - url: 'https://phishing.lacabra.app/api'
    description: 'Servidor principal'
paths:
  /blacklist:
    get:
      summary: 'Obtener la lista negra'
      description: 'Obtener la lista negra de usuarios que han sido vetados del uso de la base de datos'
      operationId: 'getBlacklist'
      tags:
        - 'Blacklist'
      security:
        - masterAuth: []
      responses:
        '200':
          description: 'Lista negra obtenida'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (200)'
                  data:
                    type: array
                    description: 'Lista negra'
                    items:
                      type: object
                      properties:
                        ip:
                          type: string
                          description: 'IP del usuario'
                        reason:
                          type: string
                          description: 'Razón del veto'
                        date:
                          type: string
                          description: 'Fecha del inicio veto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 'Añadir a la lista negra'
      description: 'Añadir a la lista negra un usuario con su IP'
      operationId: 'addBlacklist'
      tags:
        - 'Blacklist'
      security:
        - masterAuth: []
      requestBody:
        description: 'Datos del usuario a añadir'
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ip
              properties:
                ip:
                  type: string
                  description: 'IP del usuario'
                reason:
                  type: string
                  description: 'Razón del veto'
                expires_at:
                  type: string
                  description: 'Fecha de expiración del veto, en formato RFC 3339'
                created_by:
                  type: string
                  description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                notes:
                  type: string
                  description: 'Notas sobre el veto'
      responses:
        '200':
            description: 'Usuario añadido a la lista negra'
            content:
              application/json:
                schema:
                    type: object
                    properties:
                    status:
                        type: string
                        description: 'Estado de la petición (200)'
                    data:
                        type: object
                        description: 'Usuario añadido'
                        properties:
                          ip:
                            type: string
                            description: 'IP del usuario'
                          reason:
                            type: string
                            description: 'Razón del veto'
                          expires_at:
                            type: string
                            description: 'Fecha de expiración del veto, en formato RFC 3339'
                          created_at:
                            type: string
                            description: 'Fecha de creación del veto, en formato RFC 3339'
                          created_by:
                            type: string
                            description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                          notes:
                            type: string
                            description: 'Notas sobre el veto'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
            description: 'El usuario ya está en la lista negra'
            content:
                application/json:
                    schema:
                    type: object
                    properties:
                      status:
                        type: string
                        description: 'Estado de la petición (409)'
                      message:
                        type: string
                        description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
        summary: 'Modificar un veto de la lista negra'
        description: 'Modificar un veto de la lista negra con su ID de veto'
        operationId: 'patchBlacklist'
        tags:
          - 'Blacklist'
        security:
          - masterAuth: []
        parameters:
          - name: 'ip'
            in: 'path'
            description: 'IP del usuario a modificar'
            required: true
            allowEmptyValue: false
            schema:
                type:
                  - string
        requestBody:
            description: 'Datos del veto a modificar'
            required: true
            content:
                application/json:
                  schema:
                    type: object
                    properties:
                      reason:
                        type: string
                        description: 'Razón del veto'
                      expires_at:
                        type: string
                        description: 'Fecha de expiración del veto, en formato RFC 3339'
                      created_by:
                        type: string
                        description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                      notes:
                        type: string
                        description: 'Notas sobre el veto'
        responses:
            '200':
                description: 'Veto modificado'
                content:
                  application/json:
                    schema:
                    type: object
                    properties:
                        status:
                          type: string
                          description: 'Estado de la petición (200)'
                        data:
                          type: object
                          description: 'Veto modificado'
                        properties:
                            reason:
                              type: string
                              description: 'Razón del veto, solo si se ha modificado'
                            expires_at:
                              type: string
                              description: 'Fecha de expiración del veto, en formato RFC 3339, solo si se ha modificado'
                            created_at:
                              type: string
                              description: 'Fecha de creación del veto, en formato RFC 3339, solo si se ha modificado'
                            created_by:
                              type: string
                              description: 'Nombre del usuario que ha añadido al usuario a la lista negra, solo si se ha modificado'
                            notes:
                              type: string
                              description: 'Notas sobre el veto, solo si se ha modificado'
            '400':
              $ref: '#/components/responses/BadRequestError'
            '401':
              $ref: '#/components/responses/UnauthorizedError'
            '404':
                description: 'Veto no encontrado'
                content:
                  application/json:
                    schema:
                    type: object
                    properties:
                        status:
                          type: string
                          description: 'Estado de la petición (404)'
                        message:
                          type: string
                          description: 'Mensaje de error'
            '500':
              $ref: '#/components/responses/InternalServerError'

    delete:
        summary: 'Eliminar un veto de la lista negra'
        description: 'Eliminar un veto de la lista negra con su ID de veto'
        operationId: 'deleteBlacklist'
        tags:
          - 'Blacklist'
        security:
          - masterAuth: []
        parameters:
            - name: 'ip'
              in: 'path'
              description: 'IP del usuario a eliminar'
              required: true
              allowEmptyValue: false
              schema:
                type:
                  - string
        responses:
            '200':
                description: 'Veto eliminado'
                content:
                  application/json:
                    schema:
                    type: object
                    properties:
                        status:
                          type: string
                          description: 'Estado de la petición (200)'
                        data:
                          type: object
                          description: 'Veto eliminado'
                          properties:
                            ip:
                              type: string
                              description: 'IP del usuario'
                            reason:
                              type: string
                              description: 'Razón del veto'
                            expires_at:
                              type: string
                              description: 'Fecha de expiración del veto, en formato RFC 3339'
                            created_at:
                              type: string
                              description: 'Fecha de creación del veto, en formato RFC 3339'
                            created_by:
                              type: string
                              description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                            notes:
                              type: string
                              description: 'Notas sobre el veto'
            '400':
              $ref: '#/components/responses/BadRequestError'
            '401':
              $ref: '#/components/responses/UnauthorizedError'
            '404':
                description: 'Veto no encontrado'
                content:
                  application/json:
                    schema:
                    type: object
                    properties:
                        status:
                          type: string
                          description: 'Estado de la petición (404)'
                        message:
                          type: string
                          description: 'Mensaje de error'
            '500':
                $ref: '#/components/responses/InternalServerError'
  /blacklist/check:
    get:
      summary: 'Comprobar si la IP del usuario que ha hecho la petición está en la lista negra'
      description: 'Comprobar si tu IP está en la lista negra'
      operationId: 'checkBlacklist'
      tags:
        - 'Blacklist'
      responses:
        '200':
          description: 'Usuario en la lista negra'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (200)'
                  data:
                    type: object
                    description: 'Usuario en la lista negra'
                    properties:
                      ip:
                        type: string
                        description: 'IP del usuario'
                      reason:
                        type: string
                        description: 'Razón del veto'
                      expires_at:
                        type: string
                        description: 'Fecha de expiración del veto, en formato RFC 3339'
                      created_at:
                        type: string
                        description: 'Fecha de creación del veto, en formato RFC 3339'
                      created_by:
                        type: string
                        description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                      notes:
                        type: string
                        description: 'Notas sobre el veto'
        '404':
            description: 'Usuario no encontrado en la lista negra'
            content:
              application/json:
                schema:
                    type: object
                    properties:
                    status:
                        type: string
                        description: 'Estado de la petición (404)'
                    message:
                        type: string
                        description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /blacklist/{ip}:
    get:
      summary: 'Obtener un veto de la lista negra'
      description: 'Obtener un veto de la lista negra con la IP del usuario'
      operationId: 'getBlacklistIp'
      tags:
        - 'Blacklist'
      security:
        - masterAuth: []
      parameters:
        - name: 'ip'
          in: 'path'
          description: 'IP del usuario a obtener'
          required: true
          allowEmptyValue: false
          schema:
            type:
              - string
      responses:
        '200':
          description: 'Veto obtenido'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 'Estado de la petición (200)'
                  data:
                    type: object
                    description: 'Veto obtenido'
                    properties:
                      ip:
                        type: string
                        description: 'IP del usuario'
                      reason:
                        type: string
                        description: 'Razón del veto'
                      expires_at:
                        type: string
                        description: 'Fecha de expiración del veto, en formato RFC 3339'
                      created_at:
                        type: string
                        description: 'Fecha de creación del veto, en formato RFC 3339'
                      created_by:
                        type: string
                        description: 'Nombre del usuario que ha añadido al usuario a la lista negra'
                      notes:
                        type: string
                        description: 'Notas sobre el veto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
            description: 'Veto no encontrado'
            content:
              application/json:
                schema:
                    type: object
                    properties:
                    status:
                        type: string
                        description: 'Estado de la petición (404)'
                    message:
                        type: string
                        description: 'Mensaje de error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: 'Token de autenticación de usuario'
    masterAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: 'Token de autenticación de administrador, solo para uso interno'
  responses:
    UnauthorizedError:
      description: 'No autorizado'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (401)'
              message:
                type: string
                description: 'Mensaje de error'
    InternalServerError:
      description: 'Error interno del servidor'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (500)'
              message:
                type: string
                description: 'Mensaje de error'
    BadRequestError:
      description: 'Petición incorrecta o vacía'
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: 'Estado de la petición (400)'
              message:
                type: string
                description: 'Mensaje de error'
...